refactor: TimeController에서 예외처리 부분을 TimeService로 이동
